+++ docs/.vitepress/theme/demos/breadcrumb.jsx
@@ -1,0 +1,11 @@
+﻿import { Breadcrumb, BreadcrumbItem } from "../../../../src/index";
+
+export function BreadcrumbDemo() {
+  return (
+    <Breadcrumb>
+      <BreadcrumbItem href="#">Home</BreadcrumbItem>
+      <BreadcrumbItem href="#">Library</BreadcrumbItem>
+      <BreadcrumbItem>Data</BreadcrumbItem>
+    </Breadcrumb>
+  );
+}